move get_county_ids to main lib
